pyproject.toml
[tool.poetry]
name = "media_web"
version = "0.1.0"
description = ""
authors = ["RG1ee <niwows@yandex.ru>"]

packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
Pillow = "^9.1.1"
requests = "^2.28.1"
django-extensions = "^3.2.0"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
djangorestframework-simplejwt = "^5.2.0"
django-filter = "^22.1"
django-post-office = "^3.6.2"
pytest = "^7.2.0"
pytest-django = "^4.5.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
django-stubs = "^1.12.0"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
media_web = "src.manage:main"
